increase digit range for cover block difference input

+++ database/migrations/2020_02_21_102211_create_springs_table.php
@@ -17,7 +17,7 @@ class CreateSpringsTable extends Migration
             $table->bigIncrements('id');
             $table->string('color')->nullable();
             $table->string('position')->nullable();
-            $table->unsignedSmallInteger('cover_block_difference')->nullable();
+            $table->unsignedInteger('cover_block_difference')->nullable();
         });
     }
